# backend/internal/handlers/routes.go
package handlers

import (
	"net/http"
	"strings"

	"software-web-manager/backend/internal/middleware"

	"github.com/gin-gonic/gin"
)

func (h *Handler) RegisterRoutes(r *gin.Engine, installMode bool) {
	r.GET("/healthz", func(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"ok": true, "install_mode": installMode}) })
	if strings.TrimSpace(h.Cfg.LocalStoragePath) != "" {
		r.StaticFS("/files", gin.Dir(h.Cfg.LocalStoragePath, false))
	}
	r.GET("/api/ws", h.wrapWithInstallCheck(h.HandleWS))
	r.GET("/api/apps/:id/online/stream", h.wrapWithInstallCheck(h.StreamOnlineCount))

	api := r.Group("/api")
	{
		api.GET("/install/status", h.GetInstallStatus)
		api.POST("/install/test-db", h.TestDatabase)
		api.POST("/install", h.Install)
	}

	api.POST("/auth/register", h.wrapWithInstallCheck(h.Register))
	api.POST("/auth/register/send-code", h.wrapWithInstallCheck(h.SendRegisterEmailCode))
	api.POST("/auth/enterprise-register", h.wrapWithInstallCheck(h.EnterpriseRegister))
	api.GET("/auth/enterprise-status/:id", h.wrapWithInstallCheck(h.GetEnterpriseStatus))
	api.POST("/auth/enterprise-resubmit", h.wrapWithInstallCheck(h.EnterpriseResubmit))
	api.POST("/auth/login", h.wrapWithInstallCheck(h.Login))
	api.POST("/auth/admin-login", h.wrapWithInstallCheck(h.AdminLogin))
	api.POST("/auth/refresh", h.wrapWithInstallCheck(h.Refresh))
	api.GET("/public/settings", h.wrapWithInstallCheck(h.GetPublicSettings))
	api.GET("/org-invites/:token", h.wrapWithInstallCheck(h.GetOrgInvitePublic))
	api.POST("/org-invites/:token/accept", h.wrapWithInstallCheck(h.AcceptOrgInvite))

	apiAuth := r.Group("/api")
	apiAuth.Use(h.installCheckMiddleware())
	apiAuth.Use(middleware.JWT(h.Cfg))
	apiAuth.Use(h.requireJWTRequestSignature())
	apiAuth.Use(h.requireActiveUser())
	{
		apiAuth.GET("/orgs", h.ListOrgs)
		apiAuth.POST("/orgs", h.CreateOrg)
		apiAuth.GET("/orgs/:id/public", h.GetOrgPublic)
		apiAuth.POST("/orgs/:id/join-requests", h.CreateOrgJoinRequest)
		apiAuth.GET("/orgs/:id/join-requests", h.ListOrgJoinRequests)
		apiAuth.POST("/orgs/:id/join-requests/batch-delete", h.BatchDeleteOrgJoinRequests)
		apiAuth.POST("/orgs/:id/join-requests/:request_id/approve", h.ApproveOrgJoinRequest)
		apiAuth.POST("/orgs/:id/join-requests/:request_id/reject", h.RejectOrgJoinRequest)
		apiAuth.PATCH("/orgs/:id", h.UpdateOrg)
		apiAuth.POST("/orgs/upgrade", h.UpgradeOrg)
		apiAuth.POST("/orgs/:id/switch", h.SwitchOrg)
		apiAuth.POST("/orgs/:id/transfer-owner", h.TransferOwner)
		apiAuth.DELETE("/orgs/:id", h.DeleteOrg)
		apiAuth.GET("/orgs/:id/members", h.ListOrgMembers)
		apiAuth.POST("/orgs/:id/members", h.AddOrgMember)
		apiAuth.PATCH("/orgs/:id/members/:user_id", h.UpdateOrgMember)
		apiAuth.DELETE("/orgs/:id/members/:user_id", h.DeleteOrgMember)
		apiAuth.GET("/orgs/:id/roles", h.ListOrgRoles)
		apiAuth.POST("/orgs/:id/roles", h.CreateOrgRole)
		apiAuth.PATCH("/orgs/:id/roles/:role_name", h.UpdateOrgRole)
		apiAuth.DELETE("/orgs/:id/roles/:role_name", h.DeleteOrgRole)
		apiAuth.GET("/orgs/:id/permissions", h.ListOrgPermissions)
		apiAuth.GET("/orgs/:id/roles/:role_name/permissions", h.GetRolePermissions)
		apiAuth.PUT("/orgs/:id/roles/:role_name/permissions", h.PutRolePermissions)
		apiAuth.POST("/orgs/:id/users", h.CreateOrgUser)
		apiAuth.POST("/orgs/:id/invites", h.CreateOrgInvite)
		apiAuth.GET("/orgs/:id/invites", h.ListOrgInvites)
		apiAuth.DELETE("/orgs/:id/invites/:invite_id", h.RevokeOrgInvite)
		apiAuth.POST("/orgs/:id/invites/batch-delete", h.BatchDeleteOrgInvites)
		apiAuth.GET("/org-invites", h.ListUserOrgInvites)
		apiAuth.POST("/org-invites/accept-by-id/:invite_id", h.AcceptOrgInviteByID)
		apiAuth.GET("/org-join-requests/mine", h.ListMyOrgJoinRequests)
		apiAuth.POST("/org-join-requests/mine/batch-delete", h.BatchDeleteMyOrgJoinRequests)
		apiAuth.POST("/org-join-requests/mine/batch-withdraw", h.BatchWithdrawMyOrgJoinRequests)

		apiAuth.GET("/apps", h.ListApps)
		apiAuth.POST("/apps", h.CreateApp)
		apiAuth.GET("/apps/:id", h.GetApp)
		apiAuth.GET("/apps/:id/online", h.GetOnlineCount)
		apiAuth.GET("/apps/:id/online/devices", h.ListOnlineDevices)
		apiAuth.GET("/apps/:id/blocked-devices", h.ListBlockedDevices)
		apiAuth.POST("/apps/:id/blocked-devices", h.BlockDeviceByDeviceID)
		apiAuth.GET("/apps/:id/feedback", h.ListAppFeedback)
		apiAuth.GET("/apps/:id/feedback/:fid", h.GetAppFeedbackDetail)
		apiAuth.PATCH("/apps/:id/feedback/:fid", h.UpdateAppFeedback)
		apiAuth.PATCH("/apps/:id", h.UpdateApp)
		apiAuth.GET("/apps/:id/region-rules", h.GetAppRegionRules)
		apiAuth.PATCH("/apps/:id/region-rules", h.UpdateAppRegionRules)
		apiAuth.DELETE("/apps/:id", h.DeleteApp)
		apiAuth.POST("/apps/:id/channels", h.CreateChannel)
		apiAuth.GET("/apps/:id/channels", h.ListChannels)
		apiAuth.DELETE("/apps/:id/channels/:channel_id", h.DeleteChannel)
		apiAuth.GET("/apps/:id/release-channels", h.ListReleaseChannels)
		apiAuth.POST("/apps/:id/release-channels", h.CreateReleaseChannel)
		apiAuth.GET("/apps/:id/release-channels/:rcId/metrics", h.ReleaseChannelMetrics)
		apiAuth.GET("/apps/:id/members", h.ListAppMembers)
		apiAuth.POST("/apps/:id/members", h.AddAppMember)
		apiAuth.POST("/apps/:id/releases", h.CreateRelease)
		apiAuth.GET("/apps/:id/releases", h.ListReleases)
		apiAuth.PATCH("/releases/:id", h.UpdateRelease)
		apiAuth.PATCH("/releases/:id/template", h.SetReleaseTemplate)
		apiAuth.POST("/releases/:id/submit", h.SubmitRelease)
		apiAuth.POST("/releases/:id/approve", h.ApproveRelease)
		apiAuth.POST("/releases/:id/reject", h.RejectRelease)
		apiAuth.POST("/releases/:id/publish", h.PublishRelease)
		apiAuth.POST("/releases/:id/rollback", h.RollbackRelease)
		apiAuth.POST("/releases/:id/revoke", h.RevokeRelease)
		apiAuth.DELETE("/releases/:id", h.DeleteRelease)
		apiAuth.GET("/release-templates", h.ListReleaseTemplates)
		apiAuth.POST("/release-templates", h.CreateReleaseTemplate)
		apiAuth.PATCH("/release-templates/:id", h.UpdateReleaseTemplate)
		apiAuth.DELETE("/release-templates/:id", h.DeleteReleaseTemplate)
		apiAuth.PATCH("/release-channels/:id", h.UpdateReleaseChannel)
		apiAuth.POST("/releases/:id/artifacts", h.UploadArtifact)
		apiAuth.GET("/releases/:id/artifacts", h.ListArtifacts)
		apiAuth.GET("/artifacts/:id/download", h.DownloadArtifact)
		apiAuth.POST("/apps/:id/app-secrets", h.CreateAppSecret)
		apiAuth.GET("/apps/:id/app-secrets", h.ListAppSecrets)
		apiAuth.PATCH("/app-secrets/:id/policy", h.UpdateAppSecretPolicy)
		apiAuth.DELETE("/app-secrets/:id", h.RevokeAppSecret)

		apiAuth.GET("/analytics/overview", h.AnalyticsOverview)
		apiAuth.POST("/analytics/refresh", h.AnalyticsRefresh)
		apiAuth.GET("/analytics/funnel", h.AnalyticsFunnel)
		apiAuth.GET("/analytics/versions", h.AnalyticsVersions)
		apiAuth.GET("/analytics/failures", h.AnalyticsFailures)
		apiAuth.GET("/audit-logs", h.ListAuditLogs)
		apiAuth.POST("/audit-logs/delete", h.DeleteAuditLogs)
		apiAuth.GET("/tickets", h.ListTickets)
		apiAuth.GET("/tickets/:id", h.GetTicket)
		apiAuth.POST("/tickets", h.CreateTicket)
		apiAuth.PATCH("/tickets/:id/close", h.CloseTicket)
		apiAuth.DELETE("/tickets/:id", h.DeleteTicket)
		apiAuth.POST("/tickets/batch-delete", h.BatchDeleteTickets)
		apiAuth.GET("/tickets/:id/messages", h.ListTicketMessages)
		apiAuth.POST("/tickets/:id/messages", h.CreateTicketMessage)
		apiAuth.GET("/devices", h.ListDevices)
		apiAuth.POST("/devices/batch-delete", h.BatchDeleteDevices)
		apiAuth.POST("/devices/:id/block", h.BlockDevice)
		apiAuth.POST("/devices/:id/unblock", h.UnblockDevice)
		apiAuth.GET("/profile", h.GetProfile)
		apiAuth.POST("/profile/password", h.UpdateProfilePassword)
		apiAuth.POST("/profile/avatar", h.UpdateProfileAvatar)
		apiAuth.POST("/profile/2fa/setup", h.SetupProfile2FA)
		apiAuth.POST("/profile/2fa/confirm", h.ConfirmProfile2FA)
		apiAuth.POST("/profile/2fa/require", h.ToggleProfile2FA)
		apiAuth.POST("/profile/2fa/disable", h.DisableProfile2FA)
		apiAuth.GET("/geo/resolve", h.ResolveGeo)
		apiAuth.GET("/geo/regions", h.ListGeoRegions)
	}

	apiSystem := r.Group("/api/system")
	apiSystem.Use(h.installCheckMiddleware())
	apiSystem.Use(middleware.JWT(h.Cfg))
	apiSystem.Use(h.requireJWTRequestSignature())
	apiSystem.Use(h.requireActiveUser())
	apiSystem.Use(h.requireSystemAdmin())
	{
		apiSystem.GET("/overview", h.SystemOverview)
		apiSystem.GET("/orgs", h.ListSystemOrgs)
		apiSystem.GET("/orgs/:id/materials", h.ListOrgRegistrationMaterials)
		apiSystem.POST("/orgs", h.CreateSystemOrg)
		apiSystem.POST("/orgs/:id/approve", h.ApproveSystemOrg)
		apiSystem.POST("/orgs/:id/reject", h.RejectSystemOrg)
		apiSystem.POST("/orgs/:id/disable", h.DisableSystemOrg)
		apiSystem.POST("/orgs/batch-delete", h.BatchDeleteSystemOrgs)
		apiSystem.POST("/apps/batch-delete", h.BatchDeleteSystemApps)
		apiSystem.GET("/apps", h.ListSystemApps)
		apiSystem.POST("/apps/:id/approve", h.ApproveSystemApp)
		apiSystem.POST("/apps/:id/reject", h.RejectSystemApp)
		apiSystem.GET("/releases", h.ListSystemReleases)
		apiSystem.POST("/releases/:id/approve", h.ApproveSystemRelease)
		apiSystem.POST("/releases/:id/reject", h.RejectSystemRelease)
		apiSystem.GET("/audit-logs", h.ListSystemAuditLogs)
		apiSystem.POST("/audit-logs/delete", h.DeleteSystemAuditLogs)
		apiSystem.POST("/audit-logs/org-approvals/batch-delete", h.BatchDeleteSystemOrgApprovalLogs)
		apiSystem.POST("/audit-logs/app-approvals/batch-delete", h.BatchDeleteSystemAppApprovalLogs)
		apiSystem.POST("/audit-logs/release-approvals/batch-delete", h.BatchDeleteSystemReleaseApprovalLogs)
		apiSystem.GET("/tickets", h.ListSystemTickets)
		apiSystem.GET("/tickets/:id", h.GetSystemTicket)
		apiSystem.POST("/tickets", h.CreateSystemTicket)
		apiSystem.DELETE("/tickets/:id", h.DeleteSystemTicket)
		apiSystem.POST("/tickets/batch-delete", h.BatchDeleteSystemTickets)
		apiSystem.PATCH("/tickets/:id/status", h.UpdateTicketStatus)
		apiSystem.GET("/tickets/:id/messages", h.ListSystemTicketMessages)
		apiSystem.POST("/tickets/:id/messages", h.CreateSystemTicketMessage)
		apiSystem.GET("/profile", h.GetSystemProfile)
		apiSystem.POST("/profile/password", h.UpdateSystemPassword)
		apiSystem.POST("/profile/avatar", h.UpdateSystemAvatar)
		apiSystem.GET("/settings", h.GetSystemSettings)
		apiSystem.PATCH("/settings", h.UpdateSystemSettings)
		apiSystem.POST("/settings/mail/test", h.TestSystemSMTP)
		apiSystem.GET("/users", h.ListSystemUsers)
		apiSystem.PATCH("/users/:id", h.UpdateSystemUser)
		apiSystem.POST("/users/:id/enable", h.EnableSystemUser)
		apiSystem.POST("/users/:id/disable", h.DisableSystemUser)
		apiSystem.POST("/users/:id/reset-password", h.ResetSystemUserPassword)
		apiSystem.POST("/users/batch-delete", h.BatchDeleteSystemUsers)
		apiSystem.POST("/impersonate", h.Impersonate)
	}

	client := r.Group("/api/client")
	client.Use(h.installCheckMiddleware())
	client.Use(middleware.ClientLimits(h.Cfg))
	client.Use(h.requireClientSignature())
	{
		client.POST("/update-check", h.UpdateCheck)
		client.GET("/updates/stream", h.HandleClientUpdateStream)
		client.POST("/events", h.IngestEvents)
		client.POST("/feedback", h.ClientSubmitFeedback)
		client.POST("/heartbeat", h.ClientHeartbeat)
	}
}

func (h *Handler) wrapWithInstallCheck(handler gin.HandlerFunc) gin.HandlerFunc {
	return func(c *gin.Context) {
		if h.DB == nil {
			c.JSON(http.StatusServiceUnavailable, gin.H{"error": "system not installed"})
			return
		}
		handler(c)
	}
}

func (h *Handler) installCheckMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		if h.DB == nil {
			c.JSON(http.StatusServiceUnavailable, gin.H{"error": "system not installed"})
			c.Abort()
			return
		}
		c.Next()
	}
}